refactor(chat): move provider-input types to their own leaf module
Splits RawToolCallMessagePart, RawToolResultMessagePart,
ChatProviderModelInputToolCallPart, ChatProviderModelInputToolResultPart,
ChatProviderModelInputPart, and ChatProviderModelInputMessage out of
conversation.ts into provider-input-types.ts. This removes the
deliberate type-only back-edge tool-replay-reconciliation.ts had
into conversation.ts, and sets up the next increment to move
provider conversion out of conversation.ts without creating a cycle.

Pure type move, zero behaviour change.
